requirements.txt
# Python 3.8.10 compatible
Django==4.2.16
python-dotenv==1.0.1
mysqlclient==2.2.6
Pillow==10.1.0
PyPDF2==3.0.1
pdfminer.six==20221105
python-docx==1.1.0
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.0.3
numpy==1.24.4
httpx==0.25.2
